requirements.txt
# thay google-generativeai bằng google-genai
google-genai>=0.3.0

groq>=0.9.0
fastapi>=0.111.0
uvicorn>=0.30.0
langchain>=0.2.0
langchain-community>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
httpx>=0.27.0